Update the Gated-PixelCNN's input to 8 bins

+++ dqn/utils.py
@@ -7,7 +7,7 @@ from skimage.transform import resize
 imresize = resize
 
 def rgb2gray(image):
-  return np.dot(image[...,:3], [0.299, 0.587, 0.114])
+  return np.dot(image[...,:3], [0.299, 0.587, 0.114]).astype('uint8')
 
 def timeit(f):
   def timed(*args, **kwargs):
